requirements.txt
rasterio
pandas
geopandas
google-auth
google-auth-oauthlib
google-auth-httplib2
google-cloud-storage
ee